Enhancement: Split test case
