+++ test/Unit/FixtureFactory/UsingDefaultStrategyTest.php
@@ -1,0 +1,580 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2020 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/factory-bot
+ */
+
+namespace Ergebnis\FactoryBot\Test\Unit\FixtureFactory;
+
+use Ergebnis\FactoryBot\Count;
+use Ergebnis\FactoryBot\FieldDefinition;
+use Ergebnis\FactoryBot\FixtureFactory;
+use Ergebnis\FactoryBot\Test\Double;
+use Ergebnis\FactoryBot\Test\Unit;
+use Example\Entity;
+use Faker\Generator;
+
+/**
+ * @internal
+ *
+ * @covers \Ergebnis\FactoryBot\FixtureFactory
+ *
+ * @uses \Ergebnis\FactoryBot\Count
+ * @uses \Ergebnis\FactoryBot\EntityDefinition
+ * @uses \Ergebnis\FactoryBot\FieldDefinition
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\Closure
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\Optional
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\Reference
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\References
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\Sequence
+ * @uses \Ergebnis\FactoryBot\FieldDefinition\Value
+ * @uses \Ergebnis\FactoryBot\Strategy\DefaultStrategy
+ */
+final class UsingDefaultStrategyTest extends Unit\AbstractTestCase
+{
+    public function testCreateOneResolvesFieldWithoutDefaultValueToNullWhenFieldDefinitionWasNotSpecified(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            self::faker()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        self::assertNull($organization->url());
+    }
+
+    public function testCreateOneResolvesFieldWithDefaultValueToItsDefaultValueWhenFieldDefinitionWasNotSpecified(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            self::faker()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        self::assertFalse($organization->isVerified());
+    }
+
+    public function testCreateOneResolvesOneToManyAssociationToEmptyArrayCollectionWhenFieldDefinitionWasNotSpecified(): void
+    {
+        $faker = self::faker();
+
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            $faker
+        );
+
+        $fixtureFactory->define(Entity\Organization::class, [
+            'name' => $faker->word,
+        ]);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        self::assertEmpty($organization->repositories());
+    }
+
+    public function testCreateOneResolvesOptionalClosureToNullWhenFakerReturnsFalse(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::optionalClosure(static function (Generator $faker, FixtureFactory $fixtureFactory): Entity\CodeOfConduct {
+                return $fixtureFactory->createOne(Entity\CodeOfConduct::class);
+            }),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertNull($repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesOptionalClosureToResultOfClosureInvokedWithFakerAndFixtureFactoryWhenFakerReturnsTrue(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::optionalClosure(static function (Generator $faker, FixtureFactory $fixtureFactory): Entity\CodeOfConduct {
+                return $fixtureFactory->createOne(Entity\CodeOfConduct::class);
+            }),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertInstanceOf(Entity\CodeOfConduct::class, $repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesRequiredClosureToResultOfClosureInvokedWithFakerAndFixtureFactoryWhenFakerReturnsFalse(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::closure(static function (Generator $faker, FixtureFactory $fixtureFactory): Entity\CodeOfConduct {
+                return $fixtureFactory->createOne(Entity\CodeOfConduct::class);
+            }),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertInstanceOf(Entity\CodeOfConduct::class, $repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesRequiredClosureToResultOfClosureInvokedWithFakerAndFixtureFactoryWhenFakerReturnsTrue(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::closure(static function (Generator $faker, FixtureFactory $fixtureFactory): Entity\CodeOfConduct {
+                return $fixtureFactory->createOne(Entity\CodeOfConduct::class);
+            }),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertInstanceOf(Entity\CodeOfConduct::class, $repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesOptionalReferenceToNullWhenFakerReturnsFalse(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::optionalReference(Entity\CodeOfConduct::class),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertNull($repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesOptionalReferenceToEntityWhenFakerReturnsTrue(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::optionalReference(Entity\CodeOfConduct::class),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertInstanceOf(Entity\CodeOfConduct::class, $repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesRequiredReferenceToEntityWhenFakerReturnsFalse(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::reference(Entity\CodeOfConduct::class),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertInstanceOf(Entity\CodeOfConduct::class, $repository->codeOfConduct());
+    }
+
+    public function testCreateOneResolvesRequiredReferenceToEntityWhenFakerReturnsTrue(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\CodeOfConduct::class);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'codeOfConduct' => FieldDefinition::reference(Entity\CodeOfConduct::class),
+        ]);
+
+        /** @var Entity\Repository $repository */
+        $repository = $fixtureFactory->createOne(Entity\Repository::class);
+
+        self::assertInstanceOf(Entity\CodeOfConduct::class, $repository->codeOfConduct());
+    }
+
+    /**
+     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\IntProvider::greaterThanOrEqualToZero()
+     *
+     * @param int $value
+     */
+    public function testCreateOneResolvesRequiredReferencesToArrayCollectionOfEntitiesWhenFakerReturnsFalseAndCountIsExact(int $value): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class, [
+            'repositories' => FieldDefinition::references(
+                Entity\Repository::class,
+                Count::exact($value)
+            ),
+        ]);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'name' => self::faker()->word,
+        ]);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        self::assertContainsOnly(Entity\Repository::class, $organization->repositories());
+        self::assertCount($value, $organization->repositories());
+    }
+
+    /**
+     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\IntProvider::greaterThanOrEqualToZero()
+     *
+     * @param int $value
+     */
+    public function testCreateOneResolvesRequiredReferencesToArrayCollectionOfEntitiesWhenFakerReturnsTrueAndCountIsExact(int $value): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class, [
+            'repositories' => FieldDefinition::references(
+                Entity\Repository::class,
+                Count::exact($value)
+            ),
+        ]);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'name' => self::faker()->word,
+        ]);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        $repositories = $organization->repositories();
+
+        self::assertContainsOnly(Entity\Repository::class, $repositories);
+        self::assertCount($value, $repositories);
+    }
+
+    public function testCreateOneResolvesRequiredReferencesToArrayCollectionOfEntitiesWhenFakerReturnsFalseAndCountIsBetween(): void
+    {
+        $faker = self::faker();
+
+        $minimum = $faker->numberBetween(1, 5);
+        $maximum = $faker->numberBetween(10, 20);
+
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class, [
+            'repositories' => FieldDefinition::references(
+                Entity\Repository::class,
+                Count::between(
+                    $minimum,
+                    $maximum
+                )
+            ),
+        ]);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'name' => self::faker()->word,
+        ]);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        self::assertContainsOnly(Entity\Repository::class, $organization->repositories());
+        self::assertGreaterThanOrEqual($minimum, \count($organization->repositories()));
+        self::assertLessThanOrEqual($maximum, \count($organization->repositories()));
+    }
+
+    public function testCreateOneResolvesRequiredReferencesToArrayCollectionOfEntitiesWhenFakerReturnsTrueAndCountIsBetween(): void
+    {
+        $faker = self::faker();
+
+        $minimum = $faker->numberBetween(1, 5);
+        $maximum = $faker->numberBetween(10, 20);
+
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class, [
+            'repositories' => FieldDefinition::references(
+                Entity\Repository::class,
+                Count::between(
+                    $minimum,
+                    $maximum
+                )
+            ),
+        ]);
+
+        $fixtureFactory->define(Entity\Repository::class, [
+            'name' => self::faker()->word,
+        ]);
+
+        /** @var Entity\Organization $organization */
+        $organization = $fixtureFactory->createOne(Entity\Organization::class);
+
+        self::assertContainsOnly(Entity\Repository::class, $organization->repositories());
+        self::assertGreaterThanOrEqual($minimum, \count($organization->repositories()));
+        self::assertLessThanOrEqual($maximum, \count($organization->repositories()));
+    }
+
+    public function testCreateOneResolvesOptionalSequenceToNullWhenFakerReturnsFalse(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\User::class, [
+            'location' => FieldDefinition::optionalSequence('City (%d)'),
+        ]);
+
+        /** @var Entity\User $user */
+        $user = $fixtureFactory->createOne(Entity\User::class);
+
+        self::assertNull($user->location());
+    }
+
+    public function testCreateOneResolvesOptionalSequenceToStringValueWhenPercentDPlaceholderIsPresentAndFakerReturnsTrue(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\User::class, [
+            'location' => FieldDefinition::optionalSequence('City (%d)'),
+        ]);
+
+        /** @var Entity\User $userOne */
+        $userOne = $fixtureFactory->createOne(Entity\User::class);
+
+        /** @var Entity\User $userTwo */
+        $userTwo = $fixtureFactory->createOne(Entity\User::class);
+
+        /** @var Entity\User $userThree */
+        $userThree = $fixtureFactory->createOne(Entity\User::class);
+
+        self::assertSame('City (1)', $userOne->location());
+        self::assertSame('City (2)', $userTwo->location());
+        self::assertSame('City (3)', $userThree->location());
+    }
+
+    public function testCreateOneResolvesRequiredSequenceToStringValueWhenPercentDPlaceholderIsPresentAndFakerReturnsFalse(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\FalseGenerator()
+        );
+
+        $fixtureFactory->define(Entity\User::class, [
+            'location' => FieldDefinition::sequence('City (%d)'),
+        ]);
+
+        /** @var Entity\User $userOne */
+        $userOne = $fixtureFactory->createOne(Entity\User::class);
+
+        /** @var Entity\User $userTwo */
+        $userTwo = $fixtureFactory->createOne(Entity\User::class);
+
+        /** @var Entity\User $userThree */
+        $userThree = $fixtureFactory->createOne(Entity\User::class);
+
+        self::assertSame('City (1)', $userOne->location());
+        self::assertSame('City (2)', $userTwo->location());
+        self::assertSame('City (3)', $userThree->location());
+    }
+
+    public function testCreateOneResolvesRequiredSequenceToStringValueWhenPercentDPlaceholderIsPresentAndFakerReturnsTrue(): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            new Double\Faker\TrueGenerator()
+        );
+
+        $fixtureFactory->define(Entity\User::class, [
+            'location' => FieldDefinition::sequence('City (%d)'),
+        ]);
+
+        /** @var Entity\User $userOne */
+        $userOne = $fixtureFactory->createOne(Entity\User::class);
+
+        /** @var Entity\User $userTwo */
+        $userTwo = $fixtureFactory->createOne(Entity\User::class);
+
+        /** @var Entity\User $userThree */
+        $userThree = $fixtureFactory->createOne(Entity\User::class);
+
+        self::assertSame('City (1)', $userOne->location());
+        self::assertSame('City (2)', $userTwo->location());
+        self::assertSame('City (3)', $userThree->location());
+    }
+
+    /**
+     * @dataProvider \Ergebnis\FactoryBot\Test\DataProvider\IntProvider::greaterThanOrEqualToZero()
+     *
+     * @param int $value
+     */
+    public function testCreateManyResolvesToArrayOfEntitiesWhenCountIsExact(int $value): void
+    {
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            self::faker()
+        );
+
+        $fixtureFactory->define(Entity\Organization::class);
+
+        $entities = $fixtureFactory->createMany(
+            Entity\Organization::class,
+            Count::exact($value)
+        );
+
+        self::assertCount($value, $entities);
+    }
+
+    public function testCreateManyResolvesToArrayOfEntitiesWhenCountIsBetween(): void
+    {
+        $faker = self::faker();
+
+        $minimum = $faker->numberBetween(5, 10);
+        $maximum = $faker->numberBetween($minimum + 1, $minimum + 10);
+
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            $faker
+        );
+
+        $fixtureFactory->define(Entity\Organization::class);
+
+        $entities = $fixtureFactory->createMany(
+            Entity\Organization::class,
+            Count::between(
+                $minimum,
+                $maximum
+            )
+        );
+
+        self::assertGreaterThanOrEqual($minimum, \count($entities));
+        self::assertLessThanOrEqual($maximum, \count($entities));
+    }
+
+    public function testCreateManyResolvesToArrayOfEntitiesWhenFieldDefinitionOverridesAreSpecifiedAsValue(): void
+    {
+        $faker = self::faker();
+
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            $faker
+        );
+
+        $value = $faker->numberBetween(1, 5);
+
+        $fixtureFactory->define(Entity\Organization::class);
+
+        $entities = $fixtureFactory->createMany(
+            Entity\Organization::class,
+            Count::exact($value),
+            [
+                'isVerified' => true,
+            ]
+        );
+
+        self::assertCount($value, $entities);
+
+        $verifiedEntities = \array_filter($entities, static function (Entity\Organization $organization): bool {
+            return $organization->isVerified();
+        });
+
+        self::assertCount($value, $verifiedEntities);
+    }
+
+    public function testCreateManyResolvesToArrayOfEntitiesWhenFieldDefinitionOverridesAreSpecifiedAsFieldDefinition(): void
+    {
+        $faker = self::faker();
+
+        $fixtureFactory = new FixtureFactory(
+            self::entityManager(),
+            $faker
+        );
+
+        $value = $faker->numberBetween(1, 5);
+
+        $fixtureFactory->define(Entity\Organization::class);
+
+        $entities = $fixtureFactory->createMany(
+            Entity\Organization::class,
+            Count::exact($value),
+            [
+                'isVerified' => FieldDefinition::value(true),
+            ]
+        );
+
+        self::assertCount($value, $entities);
+
+        $verifiedEntities = \array_filter($entities, static function (Entity\Organization $organization): bool {
+            return $organization->isVerified();
+        });
+
+        self::assertCount($value, $verifiedEntities);
+    }
+}
